requirements-original.txt
# Core FastAPI and server dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
python-multipart>=0.0.5
python-dotenv>=0.19.0

# Database
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0

# Authentication
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AI/ML - Only for quotation processing
anthropic>=0.3.0

# PDF Processing - ESSENTIAL
PyMuPDF>=1.19.0

# OCR - EasyOCR for image processing
easyocr

# REMOVED RAG-related dependencies (moved to separate microservice):
# - openai (embeddings/LLM)
# - pinecone (vector search)  
# - llama-index packages (RAG framework)

# Image Processing
pillow==11.1.0

# Basic utilities
pydantic>=1.8.0
requests==2.32.3
python-dateutil==2.9.0.post0
# tenacity==9.0.0  # Removed - was used for RAG retry logic
